+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/text/SizedDocument.java
@@ -18,22 +18,18 @@
  */
 package is.codion.swing.common.ui.component.text;
 
+import is.codion.swing.common.ui.component.text.CaseDocumentFilter.DocumentCase;
+
 import javax.swing.text.DocumentFilter;
 import javax.swing.text.PlainDocument;
 
-/**
- * A Document implementation which allows for setting the max text length and automatic conversion to upper or lower case.
- * For instances use the {@link #sizedDocument()} or {@link #sizedDocument(int)} factory methods.
- * @see #sizedDocument()
- * @see #sizedDocument(int)
- */
-public final class SizedDocument extends PlainDocument {
+final class SizedDocument extends PlainDocument {
 
 	private final CaseDocumentFilter documentFilter;
 	private final StringLengthValidator stringLengthValidator;
 
-	private SizedDocument(int maximumLength) {
-		documentFilter = CaseDocumentFilter.caseDocumentFilter();
+	SizedDocument(int maximumLength) {
+		documentFilter = new CaseDocumentFilter(DocumentCase.NONE);
 		stringLengthValidator = new StringLengthValidator(maximumLength);
 		documentFilter.addValidator(stringLengthValidator);
 		super.setDocumentFilter(documentFilter);
@@ -48,37 +44,17 @@ public final class SizedDocument extends PlainDocument {
 		throw new UnsupportedOperationException("Changing the DocumentFilter of SizedDocument is not allowed");
 	}
 
-	@Override
-	public CaseDocumentFilter getDocumentFilter() {
-		return (CaseDocumentFilter) super.getDocumentFilter();
-	}
-
 	/**
 	 * @return the maximum length of the text to allow, -1 if unlimited
 	 */
-	public int getMaximumLength() {
+	int getMaximumLength() {
 		return stringLengthValidator.getMaximumLength();
 	}
 
 	/**
 	 * @param maximumLength the maximum length of the text to allow, -1 if unlimited
 	 */
-	public void setMaximumLength(int maximumLength) {
+	void setMaximumLength(int maximumLength) {
 		stringLengthValidator.setMaximumLength(maximumLength);
-	}
-
-	/**
-	 * @return a new {@link SizedDocument}
-	 */
-	public static SizedDocument sizedDocument() {
-		return sizedDocument(-1);
-	}
-
-	/**
-	 * @param maximumLength the maximum text length
-	 * @return a new {@link SizedDocument}
-	 */
-	public static SizedDocument sizedDocument(int maximumLength) {
-		return new SizedDocument(maximumLength);
 	}
 }