+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/text/TextFieldDocumentCase.java
@@ -29,14 +29,12 @@ final class TextFieldDocumentCase {
 	TextFieldDocumentCase(Document document, CaseDocumentFilter.DocumentCase documentCase) {
 		requireNonNull(document);
 		if (document instanceof SizedDocument) {
-			((SizedDocument) document).getDocumentFilter().setDocumentCase(documentCase);
+			((CaseDocumentFilter) ((SizedDocument) document).getDocumentFilter()).setDocumentCase(documentCase);
 		}
 		else if (document instanceof AbstractDocument) {
 			DocumentFilter documentFilter = ((AbstractDocument) document).getDocumentFilter();
 			if (documentFilter == null) {
-				CaseDocumentFilter caseDocumentFilter = CaseDocumentFilter.caseDocumentFilter();
-				caseDocumentFilter.setDocumentCase(documentCase);
-				((AbstractDocument) document).setDocumentFilter(caseDocumentFilter);
+				((AbstractDocument) document).setDocumentFilter(new CaseDocumentFilter(documentCase));
 			}
 			else if (documentFilter instanceof CaseDocumentFilter) {
 				((CaseDocumentFilter) documentFilter).setDocumentCase(documentCase);